==> Notebooks/5_RecommenderSystemWith3MostCommons.ipynb <==
# Packages importation

import networkx as nx
import glob
from collections import Counter
import pandas as pd

# Functions definition

There are 5 functions in this notebook : 
- `sorted_dict` that filter a dictionary by its values by descending order,
- `prediction_jaccard_with_edge` : It takes as arguments the viewing history of a Netflix user as a list (`ever_seen`), the list of contents that the user have not seen (`not_see`) and a graph (`graph`). For every content of `ever_seen`, the jaccard coefficient is calculated with every content of `not_see`. Each coefficient (ex : between content A and content B) is multiplicated by the weight of the link (ex : between content A and content B). Then, for each content belonging to `not_see`, we sum all recommandation score. The function returns a list of tuple ordered by decreasing score values.
- `prediction_jaccard_without_edge` : same that `prediction_jaccard_with_edge` but we don't consider the weight of the links.
- `prediction_adamic_with_edge` : same that `prediction_jaccard_with_edge` but we consider the adamic adar index instead of the jaccard coefficient.
- `prediction_adamic_without_edge` : same that `prediction_adamic_with_edge` but we don't consider the weight of the links.

def sorted_dict(d):
    return sorted(d.items(), key=lambda t: t[1], reverse=True)

def prediction_jaccard_with_edge(ever_seen, not_see, graph):
    inter = {}
    
    for elem in ever_seen:
        jac_index = [list(nx.jaccard_coefficient(graph, [(elem, i)])) for i in not_see]
        inter[elem] = {i[0][1] : i[0][2] for i in jac_index}
        
    result = {elem:[(key,value[elem]) for key, value in inter.items()] for elem in not_see}
    
    res = {key:sum([j[1]*G.get_edge_data(key,j[0])["weight"] if G.get_edge_data(key,j[0]) is not None else 0 for j in value]) for key, value in result.items()}
    
    return sorted_dict(res)

def prediction_jaccard_without_edge(ever_seen, not_see, graph):
    inter = {}
    
    for elem in ever_seen:
        jac_index = [list(nx.jaccard_coefficient(graph, [(elem, i)])) for i in not_see]
        inter[elem] = {i[0][1] : i[0][2] for i in jac_index}
        
    result = {elem:[(key,value[elem]) for key, value in inter.items()] for elem in not_see}
    
    res = {key:sum([j[1] for j in value]) for key, value in result.items()}
    
    return sorted_dict(res)

def prediction_adamic_with_edge(ever_seen, not_see, graph):
    inter = {}
    
    for elem in ever_seen:
        jac_index = [list(nx.adamic_adar_index(graph, [(elem, i)])) for i in not_see]
        inter[elem] = {i[0][1] : i[0][2] for i in jac_index}
        
    result = {elem:[(key,value[elem]) for key, value in inter.items()] for elem in not_see}
    
    res = {key:sum([j[1]*G.get_edge_data(key,j[0])["weight"] if G.get_edge_data(key,j[0]) is not None else 0 for j in value]) for key, value in result.items()}
    
    return sorted_dict(res)

def prediction_adamic_without_edge(ever_seen, not_see, graph):
    inter = {}
    
    for elem in ever_seen:
        jac_index = [list(nx.adamic_adar_index(graph, [(elem, i)])) for i in not_see]
        inter[elem] = {i[0][1] : i[0][2] for i in jac_index}
        
    result = {elem:[(key,value[elem]) for key, value in inter.items()] for elem in not_see}
    
    res = {key:sum([j[1] for j in value]) for key, value in result.items()}
    
    return sorted_dict(res)

# Data importation

- Titles

dico_id = {}
dico_id_film = {}
dico_id_serie = {}

with open("../Data/With3MostCommons/Titles.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = (line.replace('\n','')
                   .replace('\xa0', ' ')
                   .replace('*', '\*')
                   .split('\t'))
        
        dico_id[content[0]] = content[1]

with open("../Data/With3MostCommons/Titles_film.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = (line.replace('\n','')
                   .replace('\xa0', ' ')
                   .replace('*', '\*')
                   .split('\t'))
        
        dico_id_film[content[0]] = content[1]
        
with open("../Data/With3MostCommons/Titles_serie.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = (line.replace('\n','')
                   .replace('\xa0', ' ')
                   .replace('*', '\*')
                   .split('\t'))
        
        dico_id_serie[content[0]] = content[1]
# dico_title = {value:key for key,value in dico_id.items()}        
# {key:value for key,value in dico_title.items() if key.startswith('A')}

- Nodes weight

dico_node = {}
dico_node_film = {}
dico_node_serie = {}

with open("../Data/With3MostCommons/Nodes_weight.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = line.replace('\n','').split('\t')
        dico_node[content[0]] = float(content[1])

with open("../Data/With3MostCommons/Nodes_weight_film.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = line.replace('\n','').split('\t')
        dico_node_film[content[0]] = float(content[1])

with open("../Data/With3MostCommons/Nodes_weight_serie.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = line.replace('\n','').split('\t')
        dico_node_serie[content[0]] = float(content[1])
        
# dico_node

- Links

dico_link = {}
dico_link_film = {}
dico_link_serie = {}

with open("../Data/With3MostCommons/Liaisons.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = line.replace('\n','').split('\t')
        dico_link[f"{content[0]}\t{content[1]}"] = int(content[2])
        
with open("../Data/With3MostCommons/Liaisons_film.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = line.replace('\n','').split('\t')
        dico_link_film[f"{content[0]}\t{content[1]}"] = int(content[2])
        
with open("../Data/With3MostCommons/Liaisons_serie.txt","r", encoding = "utf-8") as f:
    for line in f.readlines():
        content = line.replace('\n','').split('\t')
        dico_link_serie[f"{content[0]}\t{content[1]}"] = int(content[2])
        
# dico_link

- History

files = sorted(glob.glob('../Data/Historiques_FilmsVu_Netflix/CleanHistory_user*.txt'))

dico_hist = {}
for i in range(len(files)):
    with open(files[i],"r", encoding = "utf-8") as f:
        content = f.read().split("\n")
        dico_hist[f"user{i+1}"] = content

# dico_hist

# Graph creation

G = nx.Graph()
G_film = nx.Graph()
G_serie = nx.Graph()

for key, value in dico_node.items():
    G.add_node(key,weight=value)

for key, value in dico_link.items():
    ID = key.split('\t')
    G.add_edge(ID[0], ID[1], weight=value)
    
for key, value in dico_node_film.items():
    G_film.add_node(key,weight=value)

for key, value in dico_link_film.items():
    ID = key.split('\t')
    G_film.add_edge(ID[0], ID[1], weight=value)
    
for key, value in dico_node_serie.items():
    G_serie.add_node(key,weight=value)

for key, value in dico_link_serie.items():
    ID = key.split('\t')
    G_serie.add_edge(ID[0], ID[1], weight=value)

# Recommender system

There is an interactive control here (ask the user of the notebook which user history he/she wants to analyze).

number_user = int(input(f"Entrez un numéro d'usager que vous voulez consultez (entre 1 et {len(files)}): "))

viewing_history_user = dico_hist[f"user{number_user}"]
viewing_history_user = [content for content in viewing_history_user if content in dico_id.values()]
viewing_history_user = [key for key,value in dico_id.items() if value in viewing_history_user]
content_not_see = [i for i in G.nodes() if i not in viewing_history_user]

viewing_history_user_film = dico_hist[f"user{number_user}"]
viewing_history_user_film = [content for content in viewing_history_user_film if content in dico_id_film.values()]
viewing_history_user_film = [key for key,value in dico_id_film.items() if value in viewing_history_user_film]
content_not_see_film = [i for i in G_film.nodes() if i not in viewing_history_user_film]

viewing_history_user_serie = dico_hist[f"user{number_user}"]
viewing_history_user_serie = [content for content in viewing_history_user_serie if content in dico_id_serie.values()]
viewing_history_user_serie = [key for key,value in dico_id_serie.items() if value in viewing_history_user_serie]
content_not_see_serie = [i for i in G_serie.nodes() if i not in viewing_history_user_serie]

The following chunk shows the type of movie genres that the selected user watches the most.

BDD_Netflix = pd.read_csv('../Data/BDD_Netflix.csv', sep = ';')
BDD_Netflix = BDD_Netflix.set_index('Lien Netflix')

genres_viewing_history_user = [BDD_Netflix.loc[i, "Genres Netflix"] for i in viewing_history_user]
genres_viewing_history_user = "|#|".join(genres_viewing_history_user).split("|#|")
genres_viewing_history_user = Counter(genres_viewing_history_user)
genres_viewing_history_user.most_common(15)

### Recommendation according to the Jaccard index taking into account the weight of the edges

- With the graph `G`

%%time
jac_with_edge = prediction_jaccard_with_edge(viewing_history_user, content_not_see, G)
# jac_with_edge[:10]
[(dico_id[id_title], id_title, num) for id_title,num in jac_with_edge][:5]

- With the graph `G_film`

%%time
jac_with_edge_film = prediction_jaccard_with_edge(viewing_history_user_film, content_not_see_film, G_film)
# jac_with_edge_film[:10]
[(dico_id[id_title], id_title, num) for id_title,num in jac_with_edge_film][:5]

- With the graph `G_serie`

%%time
jac_with_edge_serie = prediction_jaccard_with_edge(viewing_history_user_serie, content_not_see_serie, G_serie)
# jac_with_edge_serie[:10]
[(dico_id[id_title], id_title, num) for id_title,num in jac_with_edge_serie][:5]

### Recommendation according to the Jaccard index without taking into account the weight of the edges

- With the graph `G`

%%time
jac_without_edge = prediction_jaccard_without_edge(viewing_history_user, content_not_see, G)
[(dico_id[id_title], id_title, num) for id_title,num in jac_without_edge][:5]

- With the graph `G_film`

%%time
jac_without_edge_film = prediction_jaccard_without_edge(viewing_history_user_film, content_not_see_film, G_film)
[(dico_id[id_title], id_title, num) for id_title,num in jac_without_edge_film][:5]

- With the graph `G_serie`

%%time
jac_without_edge_serie = prediction_jaccard_without_edge(viewing_history_user_serie, content_not_see_serie, G_serie)
[(dico_id[id_title], id_title, num) for id_title,num in jac_without_edge_serie][:5]

### Recommendation according to the adamic adam index taking into account the weight of the edges

- With the graph `G`

%%time
aa_with_edge = prediction_adamic_with_edge(viewing_history_user, content_not_see, G)
[(dico_id[id_title], id_title, num) for id_title,num in aa_with_edge][:5]

- With the graph `G_film`

%%time
aa_with_edge_film = prediction_adamic_with_edge(viewing_history_user_film, content_not_see_film, G_film)
[(dico_id[id_title], id_title, num) for id_title,num in aa_with_edge_film][:5]

- With the graph `G_serie`

%%time
aa_with_edge_serie = prediction_adamic_with_edge(viewing_history_user_serie, content_not_see_serie, G_serie)
[(dico_id[id_title], id_title, num) for id_title,num in aa_with_edge_serie][:5]

### Recommendation according to the adamic adam index without taking into account the weight of the edges

- With the graph `G`

%%time
aa_without_edge = prediction_adamic_without_edge(viewing_history_user, content_not_see, G)
[(dico_id[id_title], id_title, num) for id_title,num in aa_without_edge][:5]

- With the graph `G_film`

%%time
aa_without_edge_film = prediction_adamic_without_edge(viewing_history_user_film, content_not_see_film, G_film)
[(dico_id[id_title], id_title, num) for id_title,num in aa_without_edge_film][:5]

- With the graph `G_serie`

%%time
aa_without_edge_serie = prediction_adamic_without_edge(viewing_history_user_serie, content_not_see_serie, G_serie)
[(dico_id[id_title], id_title, num) for id_title,num in aa_without_edge_serie][:5]
